requirements.txt
cloudinary
geopy
gunicorn
flask
flask-jwt-extended
flask-migrate
flask-smorest
flask-sqlalchemy
passlib
python-dotenv
psycopg2
redis
